pyproject.toml
[project]
name = "syncapp"
version = "0.1.0"
description = ""
authors = [
    {name = "Girish Reddy"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "typer (>=0.16.0,<0.17.0)",
    "selenium (>=4.33.0,<5.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "nicegui (>=2.19.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
